Keymap update & bongocat properly shut off
A lot of issues with people's BongoCat was that after shutting off the
OLED, somehow the timer would bug out & go to the next statement. Since
the animation timer keeps running, it turns on the OLED again. This
would result in the oled turning off for a few seconds, then starting up
again with idle animation.

Adding a bool to explicitly track if there is activity solves this.